+++ lib/src/threads.dart
@@ -1,4 +1,5 @@
 import 'package:chat_gpt_sdk/src/client/openai_client.dart';
+import 'package:chat_gpt_sdk/src/messages.dart';
 import 'package:chat_gpt_sdk/src/model/thread/request/thread_request.dart';
 import 'package:chat_gpt_sdk/src/model/thread/response/thread_delete_response.dart';
 import 'package:chat_gpt_sdk/src/model/thread/response/thread_response.dart';
@@ -9,13 +10,12 @@ class Threads {
 
   Threads({required OpenAIClient client}) : _client = client;
 
-  Map<String, String> _headers = {'OpenAI-Beta': 'assistants=v1'};
-  Map<String, String> get getHeader => _headers;
+  Map<String, String> get getHeader => headersAssistants;
 
   void addHeader(Map<String, String> mHeader) {
     if (mHeader == {}) return;
 
-    _headers.addAll(mHeader);
+    headersAssistants.addAll(mHeader);
   }
 
   ///Create a thread.
@@ -26,7 +26,7 @@ class Threads {
     return _client.post(
       _client.apiUrl + kThread,
       request == null ? {} : request.toJson(),
-      headers: _headers,
+      headers: headersAssistants,
       onSuccess: ThreadResponse.fromJson,
       onCancel: (cancelData) => null,
     );
@@ -39,7 +39,7 @@ class Threads {
   }) {
     return _client.get(
       _client.apiUrl + kThread + "/$threadId",
-      headers: _headers,
+      headers: headersAssistants,
       onSuccess: ThreadResponse.fromJson,
       onCancel: (cancelData) => null,
     );
@@ -52,7 +52,7 @@ class Threads {
     return _client.post(
       _client.apiUrl + kThread + "/$threadId",
       metadata,
-      headers: _headers,
+      headers: headersAssistants,
       onSuccess: ThreadResponse.fromJson,
       onCancel: (cancelData) => null,
     );
@@ -63,8 +63,13 @@ class Threads {
   }) {
     return _client.delete(
       _client.apiUrl + kThread + "/$threadId",
+      headers: headersAssistants,
       onSuccess: ThreadDeleteResponse.fromJson,
       onCancel: (cancelData) => null,
     );
   }
+
+  ///messages
+  Messages get messages =>
+      Messages(client: _client, headers: headersAssistants);
 }